+++ packages/plugin-react-refresh/index.d.ts
@@ -1,7 +1,12 @@
 import { Plugin } from 'vite'
+import { ParserOptions } from '@babel/core'
 
-type PluginFactory = () => Plugin
+type PluginFactory = (options?: Options) => Plugin
 
 declare const createPlugin: PluginFactory & { preambleCode: string }
 
-export = createPlugin
+export interface Options {
+  parserPlugins: ParserOptions['plugins']
+}
+
+export default createPlugin