+++ packages/vite/src/node/constants.ts
@@ -2,6 +2,8 @@ import path from 'path'
 
 export const SUPPORTED_EXTS = ['.mjs', '.js', '.ts', '.jsx', '.tsx', '.json']
 
+export const SPECIAL_QUERY_RE = /[\?&](worker|raw|url)\b/
+
 export const DEP_CACHE_DIR = `.vite`
 
 /**
@@ -22,11 +24,14 @@ export const VALID_ID_PREFIX = `/@id/`
 export const NULL_BYTE_PLACEHOLDER = `__x00__`
 
 export const CLIENT_PUBLIC_PATH = `/@vite/client`
+export const ENV_PUBLIC_PATH = `/@vite/env`
 // eslint-disable-next-line
 export const CLIENT_ENTRY = require.resolve('vite/dist/client/client.js')
+// eslint-disable-next-line
+export const ENV_ENTRY = require.resolve('vite/dist/client/env.js')
 export const CLIENT_DIR = path.dirname(CLIENT_ENTRY)
 
-export const knownAssetTypes = [
+export const KNOWN_ASSET_TYPES = [
   // images
   'png',
   'jpe?g',
@@ -56,7 +61,7 @@ export const knownAssetTypes = [
 ]
 
 export const DEFAULT_ASSETS_RE = new RegExp(
-  `\\.(` + knownAssetTypes.join('|') + `)(\\?.*)?$`
+  `\\.(` + KNOWN_ASSET_TYPES.join('|') + `)(\\?.*)?$`
 )
 
 export const DEP_VERSION_RE = /[\?&](v=[\w\.-]+)\b/